+++ src/main/java/com/seong/shoutlink/domain/auth/controller/AuthController.java
@@ -1,17 +1,13 @@
 package com.seong.shoutlink.domain.auth.controller;
 
-import com.seong.shoutlink.domain.auth.controller.request.CreateMemberRequest;
 import com.seong.shoutlink.domain.auth.controller.request.LoginRequest;
 import com.seong.shoutlink.domain.auth.controller.response.LoginApiResponse;
 import com.seong.shoutlink.domain.auth.service.AuthUseCase;
-import com.seong.shoutlink.domain.auth.service.request.CreateMemberCommand;
 import com.seong.shoutlink.domain.auth.service.request.LoginCommand;
-import com.seong.shoutlink.domain.auth.service.response.CreateMemberResponse;
 import com.seong.shoutlink.domain.auth.service.response.LoginResponse;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,17 +21,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class AuthController {
 
     private final AuthUseCase authUseCase;
-
-    @PostMapping("/members")
-    public ResponseEntity<CreateMemberResponse> createMember(
-        @Valid @RequestBody CreateMemberRequest request) {
-        CreateMemberResponse response = authUseCase.createMember(new CreateMemberCommand(
-            request.email(),
-            request.password(),
-            request.nickname()
-        ));
-        return ResponseEntity.status(HttpStatus.CREATED).body(response);
-    }
 
     @PostMapping("/login")
     public ResponseEntity<LoginApiResponse> login(
