--- src/main/java/com/seong/shoutlink/domain/auth/service/response/CreateMemberResponse.java
@@ -1,5 +1,0 @@
-package com.seong.shoutlink.domain.auth.service.response;
-
-public record CreateMemberResponse(Long memberId) {
-
-}
