+++ src/main/java/com/seong/shoutlink/domain/member/service/request/CreateMemberCommand.java
@@ -1,4 +1,4 @@
-package com.seong.shoutlink.domain.auth.service.request;
+package com.seong.shoutlink.domain.member.service.request;
 
 public record CreateMemberCommand(String email, String password, String nickname) {
