+++ src/main/java/com/seong/shoutlink/domain/member/service/response/CreateMemberResponse.java
@@ -1,0 +1,5 @@
+package com.seong.shoutlink.domain.member.service.response;
+
+public record CreateMemberResponse(Long memberId) {
+
+}
